Split rules into general and Fortran specific. Also split the unit tests for the Fortran rules into separate files.
